start working the functional thing of totems.

+++ common/snake/mcmods/theinvoker/net/packet/PacketTotemUpdate.java
@@ -13,7 +13,7 @@ public class PacketTotemUpdate extends PacketTI
     {
         super(PacketTypeHandler.TOTEM, true);
     }
-    public PacketTotemUpdate(int x, int y, int z, byte direction)
+    public PacketTotemUpdate(int x, int y, int z, int direction)
     {
         super(PacketTypeHandler.TOTEM, true);
         this.x = x;
@@ -25,7 +25,7 @@ public class PacketTotemUpdate extends PacketTI
     public int x;
     public int y;
     public int z;
-    public byte direction;
+    public int direction;
 
     @Override
     protected void writeData(DataOutputStream dos) throws IOException
@@ -33,7 +33,7 @@ public class PacketTotemUpdate extends PacketTI
         dos.writeInt(x);
         dos.writeInt(y);
         dos.writeInt(z);
-        dos.writeByte(direction);
+        dos.writeInt(direction);
     }
 
     @Override
@@ -42,7 +42,7 @@ public class PacketTotemUpdate extends PacketTI
         x = dis.readInt();
         y = dis.readInt();
         z = dis.readInt();
-        direction = dis.readByte();
+        direction = dis.readInt();
     }
 
     @Override
